// src/interfaces/Order/ORDER_STATUS.ts
export enum ORDER_STATUS {
  PENDING = "PENDING",
  ORDER_ACCEPTED = "ORDER_ACCEPTED",
  NEGOTIATION_PENDING = "NEGOTIATION_PENDING",
  ORDER_FINALIZED = "ORDER_FINALIZED",
  PRICE_FROZEN = "PRICE_FROZEN",
  ITEM_BOUGHT = "ITEM_BOUGHT",
  EN_ROUTE = "EN_ROUTE",
  DELIVERED = "DELIVERED",
  FUNDS_TRANSFERRED = "FUNDS_TRANSFERRED",
}